Fix double-free when >1 tabs are opened
The CURLU object for the homepage was not being duplicated when opening
a new tab. This meant that when a new document was loaded, that same
pointer was being used by each document that was initialized with
that CURLU object; this lead to a double-free when cleaning up history
for multiple tabs.